feat(stella-tui): port the SKILLS tab to the v2 SPEC 5 frame
The tab body already drew in the v2 token vocabulary; its dialogs did not,
so `views/skills.rs` was one file painting in two palettes. It moves to
`src/v2/skills.rs` with the dialogs split into `v2/skills/overlays.rs`, and
`deck_render.rs`'s `DeckTab::Skills` arm dispatches there.

The dialogs — scope picker, the LLM-assisted create flow, the pin picker,
the SKILL.md editor, the ctrl+o preview — take `stella_tui_theme::token`
in place of `crate::theme`, and the frame every one of them shares is now
the rounded rule `v2::sessions` and `v2::subagents` draw, with the keys that
dismiss it on the bottom edge rather than as a body row. Two goldens move
by exactly that much; `tab_skills.txt` does not move at all, which is what
says the body port changed nothing it drew.

Refs #4676
