+++ crates/stella-tui/README.md
@@ -154,7 +154,7 @@ escape hatch for an irreducible line (a module declaration in an oversized
 | [`src/deck_shell.rs`](src/deck_shell.rs) | `run_deck(...)`: the deck's `select!` loop, plus a third arm — a 33 ms tick that advances the clock and samples resources. |
 | [`src/envelope.rs`](src/envelope.rs) | The multi-agent wire types: `Inbound` in, `WorkspaceInput` out, and every out-of-band snapshot the driver pushes. |
 | [`src/composer.rs`](src/composer.rs) | The input model: textarea semantics, `classify_enter`, paste chips, and the slash popup shared by both surfaces. |
-| [`src/views/*.rs`](src/views) | One module per deck tab (session · agents · installed · traces · graph · files · skills · mcp · issues · settings), each exposing `render(model, ui, area, buf)`. `engine` is the exception: the config editor SETTINGS hosts, with its own `render_panel` and key handler. |
+| [`src/views/*.rs`](src/views) | One module per deck tab (session · agents · installed · traces · graph · files · mcp · issues · settings), each exposing `render(model, ui, area, buf)`. `engine` is the exception: the config editor SETTINGS hosts, with its own `render_panel` and key handler. |
 | [`src/diff.rs`](src/diff.rs), [`src/syntax.rs`](src/syntax.rs), [`src/markdown.rs`](src/markdown.rs), [`src/textline.rs`](src/textline.rs) | Shared text presentation — one implementation each of "how a diff looks", source coloring, markdown, and the event→wording table (also consumed by `stella-cli`'s plain renderer). |
 | [`src/theme.rs`](src/theme.rs), [`src/palette.rs`](src/palette.rs) | Every color and glyph. `palette.rs` mirrors the brand kit at `docs/brand/`; `theme.rs` is the only module allowed to reference it. |
 | [`src/v2/status_bar.rs`](src/v2/status_bar.rs) | SPEC 5's one-row status bar — `worker · stage · ctx [meter] % · $spend · saved $x · ✉ n`, with `? help` pinned right. `cells` is the one decision function (pure over `Status`, so the golden frames are fixture data); `render_band` draws it plus the cache-diagnosis row beneath. Replaced the two-row v1 statline in #4123/#4129; `src/statline.rs` is deleted (#4128) and this module's doc keeps the record of why the two-row shape lost. |
@@ -162,6 +162,7 @@ escape hatch for an irreducible line (a module declaration in an oversized
 | [`src/progress.rs`](src/progress.rs), [`src/cache_panel.rs`](src/cache_panel.rs), [`src/splash.rs`](src/splash.rs) | Chrome widgets: the unified stage stepper + progress row, the cache formatters behind the status bar's `saved` cell / context overlay, and the launch mark held over session init. |
 | [`src/v2/plan_card.rs`](src/v2/plan_card.rs) · [`src/v2/models_card.rs`](src/v2/models_card.rs) · [`src/v2/budget_card.rs`](src/v2/budget_card.rs) | The three floating cards, over the shared chrome in [`src/views/cards.rs`](src/views/cards.rs); their modal key handlers live in [`src/deck_ui/cards.rs`](src/deck_ui/cards.rs). `plan_card/step_style.rs` is the plan step's state→colour mapping — the card is the one surface that draws steps, now that SPEC 5 folded the always-on PLAN rail into the tab row's breadcrumb. |
 | [`src/v2/subagents.rs`](src/v2/subagents.rs) | The SUB-AGENTS overlay (`↓` from an empty prompt, `ctrl-a`, `/subagents`): every dispatched lane and every `delegate` child with its vitals (quiet time first), purpose and place, and the controls — `→`/`⏎` open, `n` nudge, `f` flag to the dispatcher, `ctrl-x ctrl-x` kill, `p` pause/resume, `r` restart, `l` back to the lead. `lifecycle.rs` is the place sentence the pulse row shares; `rows.rs` is the row model and the two messages. The one place the sub-agents are drawn; the SESSION tab stacks nothing above the transcript for them. |
+| [`src/v2/skills.rs`](src/v2/skills.rs) | The SKILLS tab (SPEC 9.2): one search box over the installed list and the npx registry, stacked so a read-aloud row carries one source. `skills/overlays.rs` holds the dialogs it floats — scope picker, the LLM-assisted create flow, the pin picker, the `SKILL.md` editor, and the `ctrl+o` markdown preview. |
 | [`src/v2/pulse.rs`](src/v2/pulse.rs) | The pulse row: the live agent's status, quiet time, place and last words, drawn in the air row above the composer on every tab but SESSION (and at an opened lane, where it is the lead). |
 | [`src/keymap.rs`](src/keymap.rs) | The one keymap table. The `?` sheet and the hint row render it; a row is a claim the witness tests under `src/deck_ui/tests/` establish. |
 | [`src/scroll.rs`](src/scroll.rs), [`src/input.rs`](src/input.rs), [`src/graph.rs`](src/graph.rs), [`src/resource.rs`](src/resource.rs), [`src/attach.rs`](src/attach.rs), [`src/clipboard.rs`](src/clipboard.rs) | Small leaf modules: line-exact viewport math, the outbound message enum, the graph snapshot types, CPU/MEM sampling, pasted-path detection, `⌃V` clipboard capture. |
