+++ crates/stella-tui/src/deck_render.rs
@@ -99,7 +99,7 @@ pub fn render_deck(model: &WorkspaceModel, ui: &mut DeckUi, frame: &mut Frame) {
         DeckTab::Traces => views::traces::render(model, ui, content, b),
         DeckTab::Graph => views::graph::render(model, ui, content, b),
         DeckTab::Files => views::files::render(model, ui, content, b),
-        DeckTab::Skills => views::skills::render(model, ui, content, b),
+        DeckTab::Skills => crate::v2::skills::render(model, ui, content, b),
         DeckTab::Mcp => views::mcp::render(model, ui, content, b),
         DeckTab::Issues => views::issues::render(model, ui, content, b),
         DeckTab::Settings => views::settings::render(model, ui, content, b),
